requirements.txt
click==7.1.1
dnspython==1.16.0
eventlet==0.25.2
Flask==1.1.2
Flask-SocketIO==4.2.1
greenlet==0.4.15
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
monotonic==1.5
pkg-resources==0.0.0
python-engineio==3.12.1
python-socketio==4.5.1
RPi.GPIO==0.7.0
six==1.14.0
Werkzeug==1.0.1
